refactor: update config imports, exports
