+++ src/config/index.js
@@ -1,6 +1,4 @@
-export { AUTH_TOKEN_EXPIRES_AT } from "./constants";
-export { REFRESH_TOKEN_EXPIRES_AT } from "./constants";
-export { BASEURL } from "./constants";
-export { default as MESSAGES } from "./messages";
-export { default as ENDPOINTS } from "./endpoints";
-export { default as URLS } from "./urls";
+export * from "./constants";
+export * from "./messages";
+export * from "./endpoints";
+export * from "./urls";
